backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4
redis[hiredis]==5.3.1
arq==0.27.0
httpx==0.28.1
pillow==12.1.1
pyjwt==2.11.0
cryptography==46.0.5
pixivpy3==3.7.5
cloudscraper==1.2.71
pydantic-settings==2.7.0
python-multipart==0.0.22
websockets==16.0
aiofiles==24.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
bcrypt==4.2.1
psutil==6.1.1
